Add files via upload
When creating visualizations for these summary tables, please be sure they make sense. You may find that creating one visualization for each table doesn't make sense because some of it may be unreadable. It is your job to determine if the visualization makes sense and if not, split it into multiple visualizations. Include a few comments
about the visualization such as "City A has the highest average texts" or something of that nature. This is the creative part of the assignment where you are expected to spend a little time creating a professional and creative solution.
For the following questions, you do not need to create visualizations but rather it will be about answering questions based on the query results. Please copy the table into the sheet for the question section (so 1 is one sheet. 2 is another sheet with A, B and C on it). To the right of the table, do analysis on the table and answer the question posed by the question. The answer may be obvious and you are encouraged to analyze it anyway, providing recommendations on how to use the data. Be creative! You still need to have these SQL queries in the SQL file. This must be done in Excel.